refactor queryForm
